:recycle: refactor components
