+++ src/features/games/components/game-of-the-day/game-of-the-day.component.html
@@ -119,6 +119,8 @@
           </div>
         </article>
       </div>
+    } @else {
+      <p>No game with those characteristics</p>
     }
   </mat-dialog-content>
