+++ src/features/games/models/index.ts
@@ -1,0 +1,1 @@
+export * from './game-card.model';
