+++ src/features/oracles/models/index.ts
@@ -1,0 +1,1 @@
+export * from './oracle-card.model';
