+++ src/features/oracles/models/oracle-card.model.ts
@@ -1,0 +1,10 @@
+import { SupportedLanguage } from '../../../shared/models';
+
+export interface OracleCard {
+  name: string;
+  artist: string;
+  language: SupportedLanguage;
+  image?: string;
+  description: string[];
+  web: string;
+}
