+++ src/shared/models/language.model.ts
@@ -1,0 +1,1 @@
+export type SupportedLanguage = 'en' | 'es' | 'de' | 'x';
